improve tests and optimize merge
